+++ opentelemetry-api/src/opentelemetry/util/re.py
@@ -32,14 +32,29 @@ _KEY_FORMAT = (
 # A value contains a URL-encoded UTF-8 string. The encoded form can contain any
 # printable US-ASCII characters (0x20-0x7f) other than SP, DEL, and ",;/
 _VALUE_FORMAT = r"[\x21\x23-\x2b\x2d-\x3a\x3c-\x5b\x5d-\x7e]*"
+# Like above with SP included
+_LIBERAL_VALUE_FORMAT = r"[\x20\x21\x23-\x2b\x2d-\x3a\x3c-\x5b\x5d-\x7e]*"
 # A key-value is key=value, with optional whitespace surrounding key and value
 _KEY_VALUE_FORMAT = rf"{_OWS}{_KEY_FORMAT}{_OWS}={_OWS}{_VALUE_FORMAT}{_OWS}"
 
 _HEADER_PATTERN = compile(_KEY_VALUE_FORMAT)
+_LIBERAL_HEADER_PATTERN = compile(
+    rf"{_OWS}{_KEY_FORMAT}{_OWS}={_OWS}{_LIBERAL_VALUE_FORMAT}{_OWS}"
+)
 _DELIMITER_PATTERN = compile(r"[ \t]*,[ \t]*")
 
 _BAGGAGE_PROPERTY_FORMAT = rf"{_KEY_VALUE_FORMAT}|{_OWS}{_KEY_FORMAT}{_OWS}"
 
+_INVALID_HEADER_ERROR_MESSAGE_STRICT_TEMPLATE = (
+    "Header format invalid! Header values in environment variables must be "
+    "URL encoded per the OpenTelemetry Protocol Exporter specification: %s"
+)
+
+_INVALID_HEADER_ERROR_MESSAGE_LIBERAL_TEMPLATE = (
+    "Header format invalid! Header values in environment variables must be "
+    "URL encoded per the OpenTelemetry Protocol Exporter specification or "
+    "a comma separated list of name=value occurrences: %s"
+)
 
 # pylint: disable=invalid-name
 
@@ -49,30 +64,51 @@ def parse_headers(s: str) -> Mapping[str, str]:
     return parse_env_headers(s)
 
 
-def parse_env_headers(s: str) -> Mapping[str, str]:
+def parse_env_headers(s: str, liberal: bool = False) -> Mapping[str, str]:
     """
     Parse ``s``, which is a ``str`` instance containing HTTP headers encoded
     for use in ENV variables per the W3C Baggage HTTP header format at
     https://www.w3.org/TR/baggage/#baggage-http-header-format, except that
     additional semi-colon delimited metadata is not supported.
+    If ``liberal`` is True we try to parse ``s`` anyway to be more compatible
+    with other languages SDKs that accept non URL-encoded headers by default.
     """
     headers: Dict[str, str] = {}
     headers_list: List[str] = split(_DELIMITER_PATTERN, s)
     for header in headers_list:
         if not header:  # empty string
             continue
-        match = _HEADER_PATTERN.fullmatch(header.strip())
-        if not match:
+        header_match = _HEADER_PATTERN.fullmatch(header.strip())
+        if not header_match and not liberal:
             _logger.warning(
-                "Header format invalid! Header values in environment variables must be "
-                "URL encoded per the OpenTelemetry Protocol Exporter specification: %s",
-                header,
+                _INVALID_HEADER_ERROR_MESSAGE_STRICT_TEMPLATE, header
             )
             continue
-        # value may contain any number of `=`
-        name, value = match.string.split("=", 1)
-        name = unquote(name).strip().lower()
-        value = unquote(value).strip()
-        headers[name] = value
+
+        if header_match:
+            match_string: str = header_match.string
+            # value may contain any number of `=`
+            name, value = match_string.split("=", 1)
+            name = unquote(name).strip().lower()
+            value = unquote(value).strip()
+            headers[name] = value
+        else:
+            # this is not url-encoded and does not match the spec but we decided to be
+            # liberal in what we accept to match other languages SDKs behaviour
+            liberal_header_match = _LIBERAL_HEADER_PATTERN.fullmatch(
+                header.strip()
+            )
+            if not liberal_header_match:
+                _logger.warning(
+                    _INVALID_HEADER_ERROR_MESSAGE_LIBERAL_TEMPLATE, header
+                )
+                continue
+
+            liberal_match_string: str = liberal_header_match.string
+            # value may contain any number of `=`
+            name, value = liberal_match_string.split("=", 1)
+            name = name.strip().lower()
+            value = value.strip()
+            headers[name] = value
 
     return headers