+++ opentelemetry-sdk/tests/metrics/exponential_histogram/test_exponential_bucket_histogram_aggregation.py
@@ -1210,14 +1210,15 @@ class TestExponentialBucketHistogramAggregation(TestCase):
         # run this test case with the same values used in a previous execution,
         # check the value printed by that previous execution of this test case
         # and use the same value for the seed variable in the line below.
-        # seed = 4539544373807492135
+        # seed = 3373389994391084876
 
         random_generator = Random(seed)
         print(f"seed for {currentframe().f_code.co_name} is {seed}")
 
         values = []
         for i in range(2000):
-            value = random_generator.randint(0, 1000)
+            # avoid both values being 0
+            value = random_generator.randint(0 if i else 1, 1000)
             values.append(value)
             histogram.aggregate(Measurement(value, now, Mock(), ctx))
             if i % 20 == 0: