+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/telemetry_attributes.py
@@ -24,11 +24,11 @@ TELEMETRY_SDK_NAME: Final = "telemetry.sdk.name"
 """
 The name of the telemetry SDK as defined above.
 Note: The OpenTelemetry SDK MUST set the `telemetry.sdk.name` attribute to `opentelemetry`.
-    If another SDK, like a fork or a vendor-provided implementation, is used, this SDK MUST set the
-    `telemetry.sdk.name` attribute to the fully-qualified class or module name of this SDK's main entry point
-    or another suitable identifier depending on the language.
-    The identifier `opentelemetry` is reserved and MUST NOT be used in this case.
-    All custom identifiers SHOULD be stable across different versions of an implementation.
+If another SDK, like a fork or a vendor-provided implementation, is used, this SDK MUST set the
+`telemetry.sdk.name` attribute to the fully-qualified class or module name of this SDK's main entry point
+or another suitable identifier depending on the language.
+The identifier `opentelemetry` is reserved and MUST NOT be used in this case.
+All custom identifiers SHOULD be stable across different versions of an implementation.
 """
 
 TELEMETRY_SDK_VERSION: Final = "telemetry.sdk.version"