+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Version 1.27.0/0.48b0 (2024-08-28)
+
+- Implementation of Events API
+  ([#4054](https://github.com/open-telemetry/opentelemetry-python/pull/4054))
+- Make log sdk add `exception.message` to logRecord for exceptions whose argument
+  is an exception not a string message
+  ([#4122](https://github.com/open-telemetry/opentelemetry-python/pull/4122))
+- Fix use of `link.attributes.dropped`, which may not exist
+  ([#4119](https://github.com/open-telemetry/opentelemetry-python/pull/4119))
 - Running mypy on SDK resources
   ([#4053](https://github.com/open-telemetry/opentelemetry-python/pull/4053))
 - Added py.typed file to top-level module
@@ -14,6 +23,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Drop Final annotation from Enum in semantic conventions
   ([#4085](https://github.com/open-telemetry/opentelemetry-python/pull/4085))
 - Update log export example to not use root logger ([#4090](https://github.com/open-telemetry/opentelemetry-python/pull/4090))
+- sdk: Add OS resource detector
+  ([#3992](https://github.com/open-telemetry/opentelemetry-python/pull/3992))
+- sdk: Accept non URL-encoded headers in `OTEL_EXPORTER_OTLP_*HEADERS` to match other languages SDKs
+  ([#4103](https://github.com/open-telemetry/opentelemetry-python/pull/4103))
+- Update semantic conventions to version 1.27.0
+  ([#4104](https://github.com/open-telemetry/opentelemetry-python/pull/4104))
+- Add support to type bytes for OTLP AnyValue
+  ([#4128](https://github.com/open-telemetry/opentelemetry-python/pull/4128))
+- Export ExponentialHistogram and ExponentialHistogramDataPoint
+  ([#4134](https://github.com/open-telemetry/opentelemetry-python/pull/4134))
+- Implement Client Key and Certificate File Support for All OTLP Exporters
+  ([#4116](https://github.com/open-telemetry/opentelemetry-python/pull/4116))
+- Remove `_start_time_unix_nano` attribute from `_ViewInstrumentMatch` in favor
+  of using `time_ns()` at the moment when the aggregation object is created
+  ([#4137](https://github.com/open-telemetry/opentelemetry-python/pull/4137))
 
 ## Version 1.26.0/0.47b0 (2024-07-25)
 
@@ -1633,4 +1657,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3778](https://github.com/open-telemetry/opentelemetry-python/pull/3778))
 - Fix license field in pyproject.toml files
   ([#3803](https://github.com/open-telemetry/opentelemetry-python/pull/3803))
-